third_party/WebKit/LayoutTests/inspector/elements/styles-4/styles-formatting-expected.txt

Tests that InspectorCSSAgent formats the CSS style text based on the CSS model modifications.

Formatted
Unformatted

Running: initFormattedStyle

Running: testFormattedInsertStart
raw cssText:
{
    /* leading comment */
    firstProperty: rgba(1, 2, 3, 0);
    property1: value1;   /* comment1 */
    property2: value2;/* comment2 */
    property3: value3;
}

Running: testFormattedRemoveStart
raw cssText:
{
    /* leading comment */
    property1: value1;   /* comment1 */
    property2: value2;/* comment2 */
    property3: value3;
}

Running: testFormattedInsertMiddle
raw cssText:
{
    /* leading comment */
    property1: value1;   /* comment1 */
    middleProperty: middleValue /* comment */;
    property2: value2;/* comment2 */
    property3: value3;
}

Running: testFormattedRemoveMiddle
raw cssText:
{
    /* leading comment */
    property1: value1;   /* comment1 */
    property2: value2;/* comment2 */
    property3: value3;
}

Running: testFormattedInsertEnd
raw cssText:
{
    /* leading comment */
    property1: value1;   /* comment1 */
    property2: value2;/* comment2 */
    property3: value3;
    endProperty: endValue;
}

Running: testFormattedRemoveEnd
raw cssText:
{
    /* leading comment */
    property1: value1;   /* comment1 */
    property2: value2;/* comment2 */
    property3: value3;
}

Running: testFormattedDisableStart
raw cssText:
{
    /* leading comment */
    /* property1: value1; */   /* comment1 */
    property2: value2;/* comment2 */
    property3: value3;
}

Running: testFormattedDisableEnd
raw cssText:
{
    /* leading comment */
    /* property1: value1; */   /* comment1 */
    property2: value2;/* comment2 */
    /* property3: value3; */
}

Running: testFormattedDisableMiddle
raw cssText:
{
    /* leading comment */
    /* property1: value1; */   /* comment1 */
    /* property2: value2; *//* comment2 */
    /* property3: value3; */
}

Running: testFormattedInsert1
raw cssText:
{
    /* leading comment */
    propA: valA;
    /* property1: value1; */   /* comment1 */
    /* property2: value2; *//* comment2 */
    /* property3: value3; */
}

Running: testFormattedInsert2
raw cssText:
{
    /* leading comment */
    propA: valA;
    /* property1: value1; */   /* comment1 */
    propB: valB;
    /* property2: value2; *//* comment2 */
    /* property3: value3; */
}

Running: testFormattedInsert3
raw cssText:
{
    /* leading comment */
    propA: valA;
    /* property1: value1; */   /* comment1 */
    propB: valB;
    /* property2: value2; *//* comment2 */
    /* property3: value3; */
    propC: valC;
}

Running: testFormattedEnableStart
raw cssText:
{
    /* leading comment */
    propA: valA;
    property1: value1;   /* comment1 */
    propB: valB;
    /* property2: value2; *//* comment2 */
    /* property3: value3; */
    propC: valC;
}

Running: testFormattedEnableEnd
raw cssText:
{
    /* leading comment */
    propA: valA;
    property1: value1;   /* comment1 */
    propB: valB;
    /* property2: value2; *//* comment2 */
    property3: value3;
    propC: valC;
}

Running: testFormattedEnableMiddle
raw cssText:
{
    /* leading comment */
    propA: valA;
    property1: value1;   /* comment1 */
    propB: valB;
    property2: value2;/* comment2 */
    property3: value3;
    propC: valC;
}

Running: testFormattedWithMeta
raw cssText:
{
    /* leading comment */
    -webkit-animation: linear;
    propA: valA;
    property1: value1;   /* comment1 */
    propB: valB;
    property2: value2;/* comment2 */
    property3: value3;
    propC: valC;
}

Running: testFormattedWithMetaValue
raw cssText:
{
    /* leading comment */
    -webkit-animation: linear;
    unicode-bidi: webkit-isolate;
    propA: valA;
    property1: value1;   /* comment1 */
    propB: valB;
    property2: value2;/* comment2 */
    property3: value3;
    propC: valC;
}

Running: testFormattedWithAtoms
raw cssText:
{
    /* leading comment */
    border-left: 1px solid rgb(1,1,1);
    -webkit-animation: linear;
    unicode-bidi: webkit-isolate;
    propA: valA;
    property1: value1;   /* comment1 */
    propB: valB;
    property2: value2;/* comment2 */
    property3: value3;
    propC: valC;
}

Running: initUnformattedStyle

Running: testUnformattedInsertStart
raw cssText:
{/*leading comment*/firstProperty: firstValue;property1:value1;property2:value2;property3:value3;}

Running: testUnformattedRemoveStart
raw cssText:
{/*leading comment*/property1:value1;property2:value2;property3:value3;}

Running: testUnformattedInsertMiddle
raw cssText:
{/*leading comment*/property1:value1;middleProperty: middleValue;property2:value2;property3:value3;}

Running: testUnformattedRemoveMiddle
raw cssText:
{/*leading comment*/property1:value1;property2:value2;property3:value3;}

Running: testUnformattedInsertEnd
raw cssText:
{/*leading comment*/property1:value1;property2:value2;property3:value3;endProperty: endValue;}

Running: testUnformattedRemoveEnd
raw cssText:
{/*leading comment*/property1:value1;property2:value2;property3:value3;}

Running: testUnformattedDisableStart
raw cssText:
{/*leading comment*//* property1:value1; */property2:value2;property3:value3;}

Running: testUnformattedDisableEnd
raw cssText:
{/*leading comment*//* property1:value1; */property2:value2;/* property3:value3; */}

Running: testUnformattedDisableMiddle
raw cssText:
{/*leading comment*//* property1:value1; *//* property2:value2; *//* property3:value3; */}

Running: testUnformattedEnableStart
raw cssText:
{/*leading comment*/property1:value1;/* property2:value2; *//* property3:value3; */}

Running: testUnformattedEnableEnd
raw cssText:
{/*leading comment*/property1:value1;/* property2:value2; */property3:value3;}

Running: testUnformattedEnableMiddle
raw cssText:
{/*leading comment*/property1:value1;property2:value2;property3:value3;}
